HW - LangChain II/deliverable/prototype/requirements.txt
pandas
python-dotenv
openai
requests
langchain-groq
pytest